Geometric Reconstruction Using Self Organizing Neural Networks (NN)
This is Visual Studio C++ code written by Jacob Barhak as part of the
Ph.D. thesis that is also provided.

The code implements several types of self organizing Neural Networks for
the purpose of reconstructing meshes from point clouds.

Some point cloud samples are provided in the samples directory.